October 29, 2024
Suppose you're given the following table, showing open and close prices as well as trading volume for a particular equity.In Python, write code to show the volume weighted average price (VWAP) for a rolling 2-day window. The VWAP is calculated as [(Price * Volume) / Volume]. In this case, your resultant table would start on 1/3/2019 and roll daily from there.