November 01, 2024
Question: Is drawing a red card and drawing an ace independent? 
Hint: Drawing an ace (event A) and drawing a red card (event B) are two separate events -- think of a way to prove (using probability theory) that knowing whether event A occurred impacts the probability of event B occuring.