November 05, 2024
Given an array of unique integers and a sum value, can you write code to find the number of triplets with a sum smaller than the given value?
We can use a sorted array and a two-pointer approach, which results in an efficient solution with a time complexity of O(n2)O(n^2)O(n2).
Solution: 
We will use two-pointer technique. Two pointers is really an easy and effective technique that is typically used for Two Sum in Sorted Arrays, Closest Two Sum, Three Sum, Four Sum, Trapping Rain Water and many other popular interview questions. Given a sorted array arr (sorted in ascending order) and a target, find if there exists any pair of elements (arr[i], arr[j]) such that their sum is equal to the target.